resolve #1 - get tests passing if MQHOST unavailable
dev.rabbitmq.com is MIA, or is so slow as to be unusable, so we
have to define MQHOST as an ENV variable. if this is not available
then we skip the tests

add Changes and Kwalitee tests, fix date format in existing Changes
file. prepare for CPAN release

perldoc all methods, add .travis.yml. travis supports rabbitmq so
hopefully we will be able to run the entire test suite on it

add MANIFEST.SKIP and complete MANIFEST, bump VERSION and Changes
for CPAN release

convert from Build.PL to Makefile.PL as CI is struggling with something
here, convert to ExtUtils::MakeMaker to try to fix these issues.
remove META.* from the repo as they will be generated as part of the
build process

we need to use Test::File::ShareDir to have the shared dir be used
correctly in testing